feat(ssh): implement SSH tunnel configuration and service
- Added `build-ssh-config.ts` to provide a utility for constructing strongly-typed SSH tunnel configurations using environment variables.
- Introduced `client.ts` to define the `SshConfig` and `SshTunnelService` classes, enabling the establishment and management of SSH tunnel connections.
- Updated `index.ts` to export the new modules, facilitating their use in other parts of the application.
